xqueue-watcher/requirements/production.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
certifi==2019.11.28
chardet==3.0.4
git+https://github.com/edx/codejail.git@1.1#egg=codejail==1.1
dogstatsd-python==0.5.6
idna==2.8
newrelic==2.98.0.81
path.py==5.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
